Split Files
Put ProjectCard into its own file
