+++ Project Portfolio V4/Project Portfolio V4.1/ClientApp/src/app/view-models/ProjectCard.ts
@@ -1,0 +1,8 @@
+export interface ProjectCard {
+  id: number;
+  name: string;
+  shortDescription: string;
+  demoUrl: string;
+  sourceUrl: string;
+  imageDataUrl: string;
+}
